+++ lib/puppet/provider/package/tdagent.rb
@@ -1,7 +1,9 @@
-# This file must be compatible with Ruby 1.8.7 in order to work on EL6.
-module Puppet::Parser::Functions
-  Puppet::Type.type(:package).provide :tdagent, parent: :gem, source: :gem do
-    has_feature :install_options, :versionable
+Puppet::Type.type(:package).provide :tdagent, parent: :gem, source: :gem do
+  has_feature :install_options, :versionable
+
+  if Puppet.features.microsoft_windows?
+    commands gemcmd: 'C:/opt/td-agent/usr/sbin/td-agent-gem'
+  else
     commands gemcmd: '/opt/td-agent/usr/sbin/td-agent-gem'
   end
 end