+++ spec/classes/install_spec.rb
@@ -6,11 +6,15 @@ RSpec.describe 'fluentd::install' do
     supported_os: [
       {
         'operatingsystem'        => 'CentOS',
-        'operatingsystemrelease' => ['6', '7'],
+        'operatingsystemrelease' => ['6', '7', '8'],
       },
       {
         'operatingsystem'        => 'Ubuntu',
-        'operatingsystemrelease' => ['16', '18'],
+        'operatingsystemrelease' => ['16', '18', '20'],
+      },
+      {
+        'operatingsystem'        => 'Windows',
+        'operatingsystemrelease' => ['2012', '2016', '2019'],
       },
     ],
   }
@@ -20,14 +24,52 @@ RSpec.describe 'fluentd::install' do
       let(:facts) { os_facts }
 
       it { is_expected.to contain_package('td-agent') }
-      it { is_expected.to contain_file('/etc/td-agent/td-agent.conf') }
-      it { is_expected.to contain_file('/etc/td-agent/config.d') }
 
-      case os_facts[:os]['family']
-      when 'RedHat'
-        it { is_expected.to contain_yumrepo('treasuredata') }
-      when 'Debian'
-        it { is_expected.to contain_apt__source('treasuredata') }
+      if os_facts[:os]['family'] == 'windows'
+        it do
+          is_expected.to contain_file('C:/opt/td-agent/etc/td-agent/td-agent.conf')
+            .with('ensure'  => 'file',
+                  'source'  => 'puppet:///modules/fluentd/td-agent.conf',
+                  'owner'   => 'Administrator',
+                  'group'   => 'Administrator',
+                  'mode'    => nil)
+        end
+        it do
+          is_expected.to contain_file('C:/opt/td-agent/etc/td-agent/config.d')
+            .with('ensure'  => 'directory',
+                  'owner'   => 'Administrator',
+                  'group'   => 'Administrator',
+                  'mode'    => nil,
+                  'recurse' => true,
+                  'force'   => true,
+                  'purge'   => true)
+        end
+      else
+        it do
+          is_expected.to contain_file('/etc/td-agent/td-agent.conf')
+            .with('ensure'  => 'file',
+                  'source'  => 'puppet:///modules/fluentd/td-agent.conf',
+                  'owner'   => 'td-agent',
+                  'group'   => 'td-agent',
+                  'mode'    => '0640')
+        end
+        it do
+          is_expected.to contain_file('/etc/td-agent/config.d')
+            .with('ensure'  => 'directory',
+                  'owner'   => 'td-agent',
+                  'group'   => 'td-agent',
+                  'mode'    => '0750',
+                  'recurse' => true,
+                  'force'   => true,
+                  'purge'   => true)
+        end
+
+        case os_facts[:os]['family']
+        when 'RedHat'
+          it { is_expected.to contain_yumrepo('treasuredata') }
+        when 'Debian'
+          it { is_expected.to contain_apt__source('treasuredata') }
+        end
       end
     end
   end