+++ spec/classes/service_spec.rb
@@ -6,11 +6,15 @@ RSpec.describe 'fluentd::service' do
     supported_os: [
       {
         'operatingsystem'        => 'CentOS',
-        'operatingsystemrelease' => ['6', '7'],
+        'operatingsystemrelease' => ['6', '7', '8'],
       },
       {
         'operatingsystem'        => 'Ubuntu',
-        'operatingsystemrelease' => ['16', '18'],
+        'operatingsystemrelease' => ['16', '18', '20'],
+      },
+      {
+        'operatingsystem'        => 'Windows',
+        'operatingsystemrelease' => ['2012', '2016', '2019'],
       },
     ],
   }
@@ -19,12 +23,7 @@ RSpec.describe 'fluentd::service' do
     context "on #{os}" do
       let(:facts) { os_facts }
 
-      case os_facts[:os]['family']
-      when 'RedHat'
-        it { is_expected.to contain_service('td-agent').with(provider: 'redhat') }
-      when 'Debian'
-        it { is_expected.to contain_service('td-agent').without(:provider) }
-      end
+      it { is_expected.to contain_service('td-agent').without(:provider) }
     end
   end
 end